flush parameter once
So a set of guessed at semantics which seem consistent are: set(value) which marks the
property as dirty, and make a note of what you need to do, then if you have a
flush, flush should do those things.

That model makes slighlty more sense to me at the resource level, but properties also
seem to work like that.