correct bad assumption
In the 'file is being created' codepath `flush` will be called, but `set` won't be.

If you leave the `set` method out you get a DevError that the provider doesn't
support `extattr=`

So, put a `set` method in, that does nothing.  Add a `flush` method that looks at
the value and runs the command.  Profit.